requirements.txt
mysql-connector-python
cryptography
pandas
openpyxl